Layers (nn): refactor QuantEltwiseAdd, add QuantCat layer
